chore: 카테고리 분석 컴포넌트 분리
