+++ src/pages/Home.vue
@@ -76,22 +76,9 @@
           </span>
         </div>
       </div>
-
       <!-- 카테고리 별 소비 분석 -->
       <div v-if="showCategory">
-        <h3>카테고리 별 소비 분석</h3>
-        <div class="progress" style="height: 3em; position: relative">
-          <div
-            v-for="(category, index) in analysisExpense()"
-            :key="index"
-            class="progress-bar"
-            role="progressbar"
-            :style="{ width: category.percentage + '%' }"
-          >
-            <span>{{ category.category }} </span>
-            <span>{{ category.percentage }}%</span>
-          </div>
-        </div>
+        <CategoryAnalysis />
       </div>
     </div>
     <button
@@ -110,6 +97,7 @@ import { useRouter } from 'vue-router';
 import axios from 'axios';
 import { useTradeListStore } from '@/stores/tradeList';
 import Calender from '@/components/Calendar.vue';
+import CategoryAnalysis from '@/components/CategoryAnalysis.vue';
 
 const nickname = ref('');
 const tradeListStore = useTradeListStore();
@@ -186,32 +174,6 @@ const sortedTradeList = (tradeList) => {
     const dateB = Number(b.tradeDate.split('-').join(''));
     return dateB - dateA;
   });
-};
-
-// 카테고리 별 소비 분석
-const analysisExpense = () => {
-  const categories = expenseList.value.map((trade) => trade.tradeMethod);
-  const categoryCounts = {};
-  categories.forEach((category) => {
-    if (categoryCounts[category]) {
-      categoryCounts[category]++;
-    } else {
-      categoryCounts[category] = 1;
-    }
-  });
-  const totalCount = categories.length;
-  const categoryPercentages = Object.entries(categoryCounts).map(
-    ([category, count]) => {
-      return {
-        category,
-        percentage: (count / totalCount) * 100,
-      };
-    }
-  );
-  categoryPercentages.forEach((category) => {
-    category.percentage = Number(category.percentage).toFixed(1);
-  });
-  return categoryPercentages;
 };
 
 const route = useRouter();
